# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/get/{name}": {
            "get": {
                "description": "Get name and index",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "get something",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Index",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetResponse"
                        }
                    }
                }
            }
        },
        "/healz": {
            "get": {
                "description": "Check health endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "healz",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Healz.response"
                        }
                    }
                }
            }
        },
        "/ready": {
            "get": {
                "description": "Check readiness",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "ready",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Ready.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.GetResponse": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer",
                    "example": 1234
                },
                "name": {
                    "type": "string",
                    "example": "test"
                }
            }
        },
        "api.Healz.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "api.Ready.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Ready"
                }
            }
        }
    }
}